[kernels]: extract shared gfx950 dualwave_swp helpers into attention
Move the DUALWAVE_SWP flash-attention primitives (ds_read transpose, s_waitcnt
encoding, exec-mask read, LDS alias scopes, split-K workspace sizing) out of
flash_attn_gfx950 into kernels/attention/dualwave_swp_common.py so the bf16/fp16
and fp8 gfx950 kernels share one copy. flash_attn_gfx950 stays a single flat
module and re-exports dualwave_splitk_workspace_elems for its public API.
